refcator: used data class in lambda instead of all parameters
