+++ cmp-navigation/src/commonMain/kotlin/cmp/navigation/navigation/FeatureNavHost.kt
@@ -18,6 +18,8 @@ import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 import org.mifos.mobile.core.common.Constants.TRANSFER_PAY_TO
 import org.mifos.mobile.core.common.FileUtils.Companion.logger
+import org.mifos.mobile.core.model.entity.AccountDetails
+import org.mifos.mobile.core.model.entity.TransferArgs
 import org.mifos.mobile.core.model.entity.TransferSuccessDestination
 import org.mifos.mobile.core.model.enums.AccountType
 import org.mifos.mobile.core.model.enums.ChargeType
@@ -59,8 +61,6 @@ import org.mifos.mobile.feature.qr.navigation.navigateToQrReaderScreen
 import org.mifos.mobile.feature.qr.navigation.qrNavGraph
 import org.mifos.mobile.feature.recent.transaction.navigation.navigateToRecentTransactionScreen
 import org.mifos.mobile.feature.recent.transaction.navigation.recentTransactionNavGraph
-import org.mifos.mobile.feature.savings.navigation.AccountDetails
-import org.mifos.mobile.feature.savings.navigation.TransferArgs
 import org.mifos.mobile.feature.savings.navigation.navigateToSavingsApplicationScreen
 import org.mifos.mobile.feature.savings.navigation.navigateToSavingsDetailScreen
 import org.mifos.mobile.feature.savings.navigation.navigateToSavingsMakeTransfer
@@ -148,18 +148,7 @@ internal fun FeatureNavHost(
             viewCharges = { chargeType, chargeTypeId ->
                 appState.navController.navigateToClientChargeScreen(chargeType, chargeTypeId)
             },
-            makePayment = { accountId, outstandingBalance, transferType, transferDestination ->
-                val args = TransferArgs(
-                    transferPayloadJson = Json.encodeToString(
-                        AccountDetails(
-                            accountId = accountId,
-                            outstandingBalance = outstandingBalance,
-                            transferType = transferType,
-                            transferTarget = TransferType.SELF,
-                            transferSuccessDestination = transferDestination,
-                        ),
-                    ),
-                )
+            makePayment = { args ->
                 appState.navController.navigateToSavingsMakeTransfer(
                     args,
                 )
@@ -264,9 +253,10 @@ fun handleHomeNavigation(
         HomeDestinations.APP_INFO -> openAppInfo()
         HomeDestinations.TRANSFER -> {
             val args = TransferArgs(
+
                 transferPayloadJson = Json.encodeToString(
                     AccountDetails(
-                        accountId = 1,
+                        accountId = 1L,
                         transferType = TRANSFER_PAY_TO,
                         transferTarget = TransferType.SELF,
                         transferSuccessDestination = TransferSuccessDestination.HOME,
