+++ feature/loan/src/commonMain/kotlin/org/mifos/mobile/feature/loan/navigation/LoanNavGraph.kt
@@ -19,7 +19,7 @@ import kotlinx.serialization.Serializable
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 import org.mifos.mobile.core.common.Constants
-import org.mifos.mobile.core.model.entity.TransferSuccessDestination
+import org.mifos.mobile.core.model.entity.TransferArgs
 import org.mifos.mobile.core.model.entity.payload.LoansPayload
 import org.mifos.mobile.core.model.enums.ChargeType
 import org.mifos.mobile.core.model.enums.LoanState
@@ -54,10 +54,7 @@ fun NavGraphBuilder.loanNavGraph(
     viewCharges: (ChargeType, Long) -> Unit,
     viewQr: (String) -> Unit,
     makePayment: (
-        accountId: Long,
-        outstandingBalance: Double?,
-        transferType: String,
-        transferSuccessDestination: TransferSuccessDestination,
+        TransferArgs,
     ) -> Unit,
 ) {
     navigation(
@@ -155,10 +152,7 @@ fun NavGraphBuilder.loanDetailRoute(
     viewTransactions: (Long) -> Unit,
     viewQr: (String) -> Unit,
     makePayment: (
-        accountId: Long,
-        outstandingBalance: Double?,
-        transferType: String,
-        transferSuccessDestination: TransferSuccessDestination,
+        TransferArgs,
     ) -> Unit,
 ) {
     composable(
